+++ crates/rustyn64-cart/src/lib.rs
@@ -22,6 +22,8 @@
 
 extern crate alloc;
 
+pub mod pi;
+
 use alloc::boxed::Box;
 use alloc::vec::Vec;